requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
pyyaml>=6.0
asyncio-mqtt>=0.11.0
aiosqlite>=0.19.0

# Exchange adapters
MetaTrader5>=5.0.0
ccxt>=4.0.0

# Technical analysis
TA-Lib>=0.4.25
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.15.0

# Notifications
python-telegram-bot>=20.0
discord-webhook>=1.3.0
smtplib-ssl>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Production
gunicorn>=21.0.0
uvicorn>=0.23.0
